# Insert new nodes at the beginning of equal range in std::unordered_multimap