- Changed the insertion and emplacement logic in `std::unordered_multimap` to insert new nodes at the beginning of an equal range instead of the back.
- This change addresses a performance issue where inserting at the back degraded performance to that of a linked list for large numbers of elements with equal keys.
- The new approach improves performance significantly, fixing a bug reported in 2013 (Bug 16747 in LLVM Bugzilla) that was previously marked as WONTFIX.
- Note: This change may break code that relies on the order of elements within each equal range, although such order is not guaranteed by the standard.
- This new behavior mimics the insertion logic of the `libstdc++` implementation.
- Added a benchmark to measure the performance improvements of the new insertion logic.