@arvidjonasson 2024-08-24
I'll look into #102303. However, I think it might be wise to initially make the fix for this issue opt-in, without waiting for #102303. Once #102303 is resolved, we can then consider making the fix opt-out. What are your thoughts on this approach?